### 2023-08-17 #5767 @dinatraykova
Reviewers suggestion:
Sbozzolo and rashti-alireza

> On 17. Aug 2023, at 19:55, The Open Journals editorial robot ***@***.***> wrote:
>
>
> Submitting author: @dinatraykova <https://github.com/dinatraykova> (Dina Traykova <http://orcid.org/0000-0002-3451-0987>)
> Repository: https://github.com/GRChombo/GRDzhadzha.git
> Branch with paper.md (empty if default branch):
> Version: v1.0
> Editor: Pending
> Reviewers: Pending
> Managing EiC: Dan Foreman-Mackey
>
> Status
>
>  <https://joss.theoj.org/papers/adcf79f964ea6c1126985118b90f06df>
> Status badge code:
>
> Author instructions
>
> Thanks for submitting your paper to JOSS @dinatraykova <https://github.com/dinatraykova>. Currently, there isn't a JOSS editor assigned to your paper.
>
> @dinatraykova <https://github.com/dinatraykova> if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search the list of people <https://reviewers.joss.theoj.org/lookup> that have already agreed to review and may be suitable for this submission.
>
> Editor instructions
>
> The JOSS submission bot @editorialbot <https://github.com/editorialbot> is here to help you find and assign reviewers and start the main review. To find out what @editorialbot <https://github.com/editorialbot> can do for you type:
>
> @editorialbot commands
