### 2023-10-17 #5956 @editorialbot
[REVIEW]: GRDzhadzha: A code for evolving relativistic matter on analytic metric backgrounds
**Submitting author:** @dinatraykova (<a href="http://orcid.org/0000-0002-3451-0987">Dina Traykova</a>)
**Repository:** https://github.com/GRChombo/GRDzhadzha.git
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
