### 2023-11-16 #5956 @rashti-alireza
@dinatraykova :
Regarding the "Contribution and authorship" checklist, as I examined the git log of the repository, I observed that some authors listed in the paper do not have any associated commits. Furthermore, based on the git log, it appears that certain authors solely made changes to text and file names. Could you please provide clarification on these specific points? Thank you!
