### 2023-11-22 #5956 @ekwessel
I have completed my review, and as far as I can see, everything looks good!

The biggest hold-up for me was figuring out how to install GRChombo (which I have never used before) on a new cluster. The process involved some trial and error, drawing on a good amount of experience configuring codes in HPC environments. However, for performant NR codes such as this, cluster-specific configuration is essential, and it is reasonable to expect any potential users ought to have the necessary technical skills. Additionally, I found the GRChombo documentation provided very helpful guidance. In particular the repository of common issues and their resolutions was crucial to enabling me to identify and fix issues with my configuration. So, while this slowed me down, I don't believe it is a issue. After GRChombo was installed and the tests were passing, installing GRDzhadzha based on the provided instructions in this repository was easy.

I built and ran the tests, which passed, as well as the examples, and was able to plot their output, which looks sensible. The only statement about performance in the paper involved a very rough comparison between comparable situations in GRChombo and GRDzhadzha: I was able to compare the runtime of one of the GRChombo examples to the similar BoostedBH example here, and can verify that the claims about the relative speedup are accurate.

I am therefore recommending acceptance of this code. Well done!
