### 2023-11-22 #5956 @ekwessel
@rashti-alireza I do have one small question: what inspired the name "GRDzhadzha"?
