### 2023-11-22 #5956 @rashti-alireza
@dinatraykova
I'm at the software paper checklist.
The paper demonstrates a high standard of writing and aligns mostly with the criteria set by the JOSS journal. I do have a few points to be clarified to proceed. Please find the raised points in the following section:

- [ ]  Please provide citations for the sentence or statement takes place in the following lines:

- line 17:"study questions in fundamental physics, such as the existence and properties..."
- line 19:"...one imposes significant symmetries"
- line 26:"...in many cases of interest..."
- line 27:"in which case it is a reasonable approximation..."
- line 47:"...can easily be adapted to other ..."
- line 50:"an ADM decomposition..."
- line 59:"extrapolating"
- line 59:"Sommerfeld"

- [ ] Please be specific about what you mean by "significantly speeds up ..." at the line 34.
- [ ]  The figure doesn't have any caption, please provide a descriptive caption and a citation.
- [ ]  There is no actual citation link at the line 130.
- [ ] There are sections that are called "Key features of GRChombo" and "Key research projects using GRChombo"
  why "GRChombo"? and not "GRDzhadzha"?
