### 2023-11-22 #5956 @rashti-alireza
@dinatraykova
Since the code relies on analytically derived metrics , please provide me convergence testings of the Hamiltonian and momentum constraints for a black hole in the Kerr-Schild coordinates and a boosted black hole in
isotropic Schwarzschild coordinates. Additionally, please ensure these tests conform to the finite difference scheme employed in the code.
