### 2023-11-22 #5956 @rashti-alireza
@dinatraykova
Can you explain what the 'Automated tests' are? What do they actually test?
