+++ paper/paper.bib
@@ -921,6 +921,95 @@
     year = "1916"
 }
 
+@article{Barausse:2014tra,
+    author = "Barausse, Enrico and Cardoso, Vitor and Pani, Paolo",
+    title = "{Can environmental effects spoil precision gravitational-wave astrophysics?}",
+    eprint = "1404.7149",
+    archivePrefix = "arXiv",
+    primaryClass = "gr-qc",
+    doi = "10.1103/PhysRevD.89.104059",
+    journal = "Phys. Rev. D",
+    volume = "89",
+    number = "10",
+    pages = "104059",
+    year = "2014"
+}
+
+@article{Barausse:2014pra,
+    author = "Barausse, Enrico and Cardoso, Vitor and Pani, Paolo",
+    editor = "Ciani, Giacomo and Conklin, John W. and Mueller, Guido",
+    title = "{Environmental Effects for Gravitational-wave Astrophysics}",
+    eprint = "1404.7140",
+    archivePrefix = "arXiv",
+    primaryClass = "astro-ph.CO",
+    doi = "10.1088/1742-6596/610/1/012044",
+    journal = "J. Phys. Conf. Ser.",
+    volume = "610",
+    number = "1",
+    pages = "012044",
+    year = "2015"
+}
+
+@article{Barack:2018yly,
+    author = "Barack, Leor and others",
+    title = "{Black holes, gravitational waves and fundamental physics: a roadmap}",
+    eprint = "1806.05195",
+    archivePrefix = "arXiv",
+    primaryClass = "gr-qc",
+    doi = "10.1088/1361-6382/ab0587",
+    journal = "Class. Quant. Grav.",
+    volume = "36",
+    number = "14",
+    pages = "143001",
+    year = "2019"
+}
+
+@article{Macedo:2013qea,
+    author = "Macedo, Caio F. B. and Pani, Paolo and Cardoso, Vitor and Crispino, Lu\'\i{}s C. B.",
+    title = "{Into the lair: gravitational-wave signatures of dark matter}",
+    eprint = "1302.2646",
+    archivePrefix = "arXiv",
+    primaryClass = "gr-qc",
+    doi = "10.1088/0004-637X/774/1/48",
+    journal = "Astrophys. J.",
+    volume = "774",
+    pages = "48",
+    year = "2013"
+}
+
+@article{Bertone:2019irm,
+    author = "Bertone, Gianfranco and others",
+    title = "{Gravitational wave probes of dark matter: challenges and opportunities}",
+    eprint = "1907.10610",
+    archivePrefix = "arXiv",
+    primaryClass = "astro-ph.CO",
+    doi = "10.21468/SciPostPhysCore.3.2.007",
+    journal = "SciPost Phys. Core",
+    volume = "3",
+    pages = "007",
+    year = "2020"
+}
+
+@article{Arnowitt:1962hi,
+    author = "Arnowitt, Richard L. and Deser, Stanley and Misner, Charles W.",
+    title = "{The Dynamics of general relativity}",
+    eprint = "gr-qc/0405109",
+    archivePrefix = "arXiv",
+    doi = "10.1007/s10714-008-0661-1",
+    journal = "Gen. Rel. Grav.",
+    volume = "40",
+    pages = "1997--2027",
+    year = "2008"
+}
+
+@inproceedings{York:1978gql,
+    author = "York, Jr., James W.",
+    title = "{Kinematics and Dynamics of General Relativity}",
+    booktitle = "{Workshop on Sources of Gravitational Radiation}",
+    pages = "83--126",
+    year = "1978"
+}
+
 @article{Traykova:2021dua,
[…]
