+++ paper/paper.bib
@@ -1010,6 +1010,29 @@
     year = "1978"
 }
 
+@article{Sommerfeld1912,
+author = {Sommerfeld, A.},
+journal = {Jahresbericht der Deutschen Mathematiker-Vereinigung},
+pages = {309-352},
+title = {Die Greensche Funktion der Schwingungslgleichung.},
+url = {http://eudml.org/doc/145344},
+volume = {21},
+year = {1912},
+}
+
+@article{Alcubierre:2002kk,
+    author = "Alcubierre, Miguel and Bruegmann, Bernd and Diener, Peter and Koppitz, Michael and Pollney, Denis and Seidel, Edward and Takahashi, Ryoji",
+    title = "{Gauge conditions for long term numerical black hole evolutions without excision}",
+    eprint = "gr-qc/0206072",
+    archivePrefix = "arXiv",
+    reportNumber = "AEI-2002-037",
+    doi = "10.1103/PhysRevD.67.084023",
+    journal = "Phys. Rev. D",
+    volume = "67",
+    pages = "084023",
+    year = "2003"
+}
+
 @article{Traykova:2021dua,
     author = "Traykova, Dina and Clough, Katy and Helfer, Thomas and Berti, Emanuele and Ferreira, Pedro G. and Hui, Lam",
     title = "{Dynamical friction from scalar dark matter in the relativistic regime}",
@@ -1106,11 +1129,8 @@
 @article{Wang:inprep,
     author = "Wang, Zipeng and Helfer, Thomas and Traykova, Dina and Clough, Katy and Berti, Emanuele",
     title = "{The Magnus Effect on black holes in scalar dark matter environments}",
-    eprint = "in preparation",
-    archivePrefix = "arXiv",
-    primaryClass = "gr-qc",
     doi = "",
-    journal = "",
+    journal = "in preparation",
     volume = "",
     number = "",
     pages = "",
