+++ paper/paper.md
@@ -45,7 +45,7 @@ affiliations:
   index: 3
 - name: Department of Applied Mathematics and Theoretical Physics (DAMTP), University of Cambridge, Centre for Mathematical Sciences, Wilberforce Road, Cambridge CB3 0WA, United Kingdom
   index: 4
-- name: Max Planck Institute for Gravitational Physics (Albert Einstein Institute), Am M\"uhlenberg 1, Potsdam-Golm, 14476, Germany
+- name: Max Planck Institute for Gravitational Physics (Albert Einstein Institute), Am Mühlenberg 1, Potsdam-Golm, 14476, Germany
   index: 5
 date: 14 Aug 2023
 bibliography: paper.bib
@@ -54,7 +54,7 @@ bibliography: paper.bib
 
 # Summary
 
-Strong gravity environments such as those around black holes provide us with unique opportunities to study questions in fundamental physics, such as the existence and properties of dark matter and dark energy. Characterising the behaviour of new fields and other types of matter in highly relativistic environments generally necessitates numerical simulations unless one imposes significant symmetries.
+Strong gravity environments such as those around black holes provide us with unique opportunities to study questions in fundamental physics (see e.g [@Macedo:2013qea;@Barausse:2014pra;@Barack:2018yly;@Bertone:2019irm]), such as the existence and properties of dark matter and dark energy. Characterising the behaviour of new fields and other types of matter in highly relativistic environments generally necessitates numerical simulations unless one imposes significant symmetries.
 Therefore we need to turn to numerical methods to study the dynamics and evolution of the complex systems of black holes and other compact objects in different environments, using numerical relativity (NR).
 These methods allow us to split the four-dimensional Einstein equations into three-dimensional spatial hypersurfaces and a time-like direction.
 Then if a solution is known at the initial spatial hypersurface, it can be numerically evolved in time, where an analytic solution no longer exists.
@@ -73,10 +73,10 @@ GRDzhadzha inherits many of the features of GRChombo and Chombo, but avoids the 
 
 - Background metrics: The currently available backgrounds in the code are a static Kerr black hole in horizon penetrating Kerr-Schild coordinates and a boosted black hole in isotropic Schwarzschild coordinates. These backgrounds can easily be adapted to other coordinate systems for different problems. The code is templated over the background so it can easily be changed without major code modification.
 
-- Matter evolution: The code calculates the evolution for the matter variables on the metric background using an ADM decomposition in space and time - currently we have implemented a real and a complex scalar field as examples of matter types. Again the code is templated over the matter class so that the matter types can be exchanged with minimal modification.
+- Matter evolution: The code calculates the evolution for the matter variables on the metric background using an ADM decomposition [@Arnowitt:1962hi;@York:1978gql] in space and time - currently we have implemented a real and a complex scalar field as examples of matter types. Again the code is templated over the matter class so that the matter types can be exchanged with minimal modification.
 
 - Accuracy: The metric values and their derivatives are calculated exactly at each point, whereas the matter fields are evolved with a 4th order Runge-Kutta time integration and their derivatives calculated with the same finite difference stencils used in GRChombo (4th and 6th order are currently available).
-- Boundary Conditions: GRDzhadzha inherits all the available boundary conditions in GRChombo, namely, extrapolating, Sommerfeld (radiative), reflective and periodic. 
+- Boundary Conditions: GRDzhadzha inherits all the available boundary conditions in GRChombo, namely, extrapolating, Sommerfeld (radiative) [@Sommerfeld1912], reflective and periodic. 
[…]
