+++ paper/paper.md
@@ -54,11 +54,11 @@ bibliography: paper.bib
 
 # Summary
 
-Strong gravity environments such as those around black holes provide us with unique opportunities to study questions in fundamental physics (see e.g [@Macedo:2013qea;@Barausse:2014pra;@Barack:2018yly;@Bertone:2019irm]), such as the existence and properties of dark matter and dark energy. Characterising the behaviour of new fields and other types of matter in highly relativistic environments generally necessitates numerical simulations unless one imposes significant symmetries.
+Strong gravity environments such as those around black holes provide us with unique opportunities to study questions in fundamental physics (see e.g [@Macedo:2013qea;@Barausse:2014pra;@Barack:2018yly;@Bertone:2019irm]), such as the existence and properties of dark matter and dark energy. Characterising the behaviour of new fields and other types of matter in highly relativistic environments usually necessitates numerical simulations unless one imposes significant symmetries.
 Therefore we need to turn to numerical methods to study the dynamics and evolution of the complex systems of black holes and other compact objects in different environments, using numerical relativity (NR).
 These methods allow us to split the four-dimensional Einstein equations into three-dimensional spatial hypersurfaces and a time-like direction.
 Then if a solution is known at the initial spatial hypersurface, it can be numerically evolved in time, where an analytic solution no longer exists.
-Whilst the tools of NR provide the most complete (i.e., approximation free) method for evolving matter in such environments, in many cases of interest, the density of the matter components is negligible in comparison to the curvature scales of the background spacetime metric, in which case it is a reasonable approximation to neglect their backreaction on it and treat the metric as fixed (assuming the background itself is stationary or otherwise has an analytic form).
+Whilst the tools of NR provide the most complete (i.e., approximation free) method for evolving matter in such environments, in many cases of interest, the density of the matter components is negligible in comparison to the curvature scales of the background spacetime metric [@Clough:2021qlv]. In such cases it is a reasonable approximation to neglect the backreaction of the matter environment onto the metric and treat it as fixed (assuming the background itself is stationary or otherwise has an analytic form).
 
 In such cases, one does not need to evolve all the metric degrees of freedom as in NR, but only the additional matter ones. 
 It is possible to do this using any NR code in a trivial way by setting the evolution of the metric variables to zero, but this is clearly rather inefficient. This code, GRDzhadzha, directly evolves the matter variables on an analytically specified background.
