### 2023-12-01 64e08d0
Update run-grdzhadzha-tests-intel.yml
