### 2024-02-06 #5956 @rashti-alireza
> > > > @dinatraykova
> > > > Since the code relies on analytically derived metrics , please provide me convergence testings of the Hamiltonian and momentum constraints for a black hole in the Kerr-Schild coordinates and a boosted black hole in
> > > > isotropic Schwarzschild coordinates. Additionally, please ensure these tests conform to the finite difference scheme employed in the code. Ideally, I need a plot showing the constraints are decreased by increasing the resolution with a proper slope.
> > >
> > >
> > > @rashti-alireza This is what the tests in the Test folder do. For each example, they show that the time derivatives of the metric functions, and the Hamiltonian and momentum constraints for the initial data are really zero (see Tests/BoostedBHComplexScalarTest/BoostedBHScalarTest.cpp and Tests/KerrBHScalarTest/KerrBHScalarTest.cpp)
> >
> >
> > I would like to see a convergence test for the Hamiltonian constraint of a Kerr-Schild BH, i.e., the quantity H:=R−KijKij+K2. I don't expect this quantity be "really zero" on a finite grid if R and Kij are constructed from the Kerr-Schild metric. In this case H should only converge to zero (unless you also compute R and Kij analytically). Thanks!
>
> @rashti-alireza Maybe my previous response wasn't very clear, the convergence tests that we do show that the norm of the constraints converges to zero (see line 230 here: https://github.com/GRTLCollaboration/GRDzhadzha/blob/main/Tests/KerrBHScalarTest/KerrBHScalarTest.cpp )
>
> Hope this clears things up :)

@dinatraykova, Great! that answers my question now.
