### 2024-02-06 #5956 @rashti-alireza
> Hi @rashti-alireza, does the convergence test that @dinatraykova linked satisfy you? I know little about numerical GR but I can at least see that [the code snippet](https://github.com/GRTLCollaboration/GRDzhadzha/blob/b5f430c7896e9f1a1999dbedc89368f06bf6320a/Tests/KerrBHScalarTest/KerrBHScalarTest.cpp#L230-L262) is a convergence test. I might be misunderstanding the jargon but is it an issue for you that this is a convergence test on the norm of the constraints, rather than a convergence test of the single constraint that you highlighted [above](https://github.com/openjournals/joss-reviews/issues/5956#issuecomment-1879182827)?
>
> As an aside for @dinatraykova, I just a look myself that the tests are indeed automated but noticed that the `ifort` build started failing about four months ago.

Hi @warrickball, everything is clear now and all of my questions have been answered adequately.
The authors did a great work developing the GRDzhadzha code. The code has great potentials and I highly recommend it for the publication. Great job!
