### 2024-02-07 d8848ea
minor editorial changes for JOSS
