### 2024-02-07 #5956 @warrickball
Hi again @dinatraykova, I've [proposed a few small changes](https://github.com/GRTLCollaboration/GRDzhadzha/pull/20) to the paper but a few more changes are required that I'll tick off as we resolve.

* [ ] There are several references that have authors "Someone & others". Could you update the BibTeX entries to use the full author lists? NASA ADS should generate good BibTeX entries.
* [ ] For Chombo, it'd be nice to have a less opaque reference (i.e., something people can realistically find) but I accept if this isn't possible.
* [ ] The figure caption isn't rendering correctly. I think what you need to do is put the two images together into a single file and then use the standard markdown figure captioning (i.e., something like `![Caption](image.png)`). See the [JOSS docs](https://joss.readthedocs.io/en/latest/submitting.html#figures-and-images).
* [ ] Many citations are bracketed but shouldn't be. e.g. in the first sentence
> (see e.g (Barack & others, 2019; Barausse et al., 2015; Bertone & others, 2020; Macedo et al., 2013))

should be
> (see e.g Barack et al., 2019; Barausse et al., 2015; Bertone et al., 2020; Macedo et al., 2013)

I remember this being tricky so let me know if you have trouble. See the [JOSS docs](https://joss.readthedocs.io/en/latest/submitting.html#citations) and [pandoc docs](https://pandoc.org/MANUAL#extension-citations) for the citation options.
* [ ] You have lots of inline URLs that [should be rendered as hyperlinks](https://joss.readthedocs.io/en/latest/submitting.html#links). This should be straightforward: just collapse e.g. in the markdown source, `Einstein Toolkit (http://einsteintoolkit.org/)` should be `[Einstein Toolkit](http://einsteintoolkit.org/)`.
