### 2024-03-04 6c94921
Update paper.bib with full author lists
