+++ paper/paper.bib
@@ -1,16 +1,23 @@
-@article{Andrade:2021rbd,
-    author = "Andrade, Tomas and others",
-    title = "{GRChombo: An adaptable numerical relativity code for fundamental physics}",
-    eprint = "2201.03458",
-    archivePrefix = "arXiv",
-    primaryClass = "gr-qc",
-    doi = "10.21105/joss.03703",
-    journal = "J. Open Source Softw.",
-    volume = "6",
-    number = "68",
-    pages = "3703",
-    year = "2021"
+@ARTICLE{Andrade:2021rbd,
+       author = {{Andrade}, Tomas and {Salo}, Llibert and {Aurrekoetxea}, Josu and {Bamber}, Jamie and {Clough}, Katy and {Croft}, Robin and {de Jong}, Eloy and {Drew}, Amelia and {Duran}, Alejandro and {Ferreira}, Pedro and {Figueras}, Pau and {Finkel}, Hal and {Fran{\c{c}}a}, Tiago and {Ge}, Bo-Xuan and {Gu}, Chenxia and {Helfer}, Thomas and {J{\"a}ykk{\"a}}, Juha and {Joana}, Cristian and {Kunesch}, Markus and {Kornet}, Kacper and {Lim}, Eugene and {Muia}, Francesco and {Nazari}, Zainab and {Radia}, Miren and {Ripley}, Justin and {Shellard}, Paul and {Sperhake}, Ulrich and {Traykova}, Dina and {Tunyasuvunakool}, Saran and {Wang}, Zipeng and {Widdicombe}, James and {Wong}, Kaze},
+        title = "{GRChombo: An adaptable numerical relativity code for fundamental physics}",
+      journal = {The Journal of Open Source Software},
+     keywords = {numerical relativity, Open MP, vector intrinsics, general relativity, Mathematica, gravity, c++, MPI, C++, General Relativity and Quantum Cosmology},
+         year = 2021,
+        month = dec,
+       volume = {6},
+       number = {68},
+          eid = {3703},
+        pages = {3703},
+          doi = {10.21105/joss.03703},
+archivePrefix = {arXiv},
+       eprint = {2201.03458},
+ primaryClass = {gr-qc},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2021JOSS....6.3703A},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+
 
 @article{Radia:2021hjs,
     author = "Radia, Miren and Sperhake, Ulrich and Berti, Emanuele and Croft, Robin",
@@ -96,19 +103,23 @@
     year = "2022"
 }
 
-@article{LIGOScientific:2020ibl,
-    author = "Abbott, R. and others",
-    collaboration = "LIGO Scientific, Virgo",
-    title = "{GWTC-2: Compact Binary Coalescences Observed by LIGO and Virgo During the First Half of the Third Observing Run}",
-    eprint = "2010.14527",
-    archivePrefix = "arXiv",
-    primaryClass = "gr-qc",
-    reportNumber = "P2000061",
-    doi = "10.1103/PhysRevX.11.021053",
[…]
