### 2024-03-04 6cf29ff
Update paper.md
Fixed the figure caption, references and urls display.
