### 2024-03-04 43761fa
Update paper.bib
Fix typo
