+++ paper/paper.bib
@@ -612,7 +612,7 @@ archivePrefix = {arXiv},
     title = "{Chombo software package for AMR applications - design document}",
     reportNumber = "LBNL-6616E",
     month = "12",
-    year = "2015"
+    year = "2015",
     url= "https://commons.lbl.gov/display/chombo"
 }
 
