### 2024-03-04 #5956 @dinatraykova
Thanks again @rashti-alireza and @ekwessel for your comments and suggestions!
And thanks @warrickball for you editorial help!
I've addressed your comments above in this pull request https://github.com/GRTLCollaboration/GRDzhadzha/pull/21
For the Chombo reference, I added a url and the full author list, which I think makes it easier to find.

Regarding the failing intel test, that started happening recently due to some changes in the main GRChombo code, which someone is working on resolving.
