### 2024-03-06 a5ffcad
Add boosted bh proca example
