+++ Examples/BoostedBHProca/BoostedBHProcaLevel.cpp
@@ -1,0 +1,164 @@
+/* GRChombo
+ * Copyright 2012 The GRChombo collaboration.
+ * Please refer to LICENSE in GRChombo's root directory.
+ */
+
+// General includes common to most GR problems
+#include "BoostedBHProcaLevel.hpp"
+#include "AMRReductions.hpp"
+#include "BoxLoops.hpp"
+#include "ComputePack.hpp"
+#include "NanCheck.hpp"
+#include "SetValue.hpp"
+#include "SmallDataIO.hpp"
+
+// For RHS update
+#include "BoostedBH.hpp"
+#include "MatterEvolution.hpp"
+
+// For tag cells
+#include "FixedGridsTaggingCriterion.hpp"
+
+// Problem specific includes
+#include "EnergyConservation.hpp"
+#include "ExcisionDiagnostics.hpp"
+#include "ExcisionEvolution.hpp"
+#include "FluxExtraction.hpp"
+#include "InitialProcaData.hpp"
+#include "LinearMomConservation.hpp"
+#include "ProcaConstraint.hpp"
+#include "ProcaField.hpp"
+
+// Initial data for field and metric variables
+void BoostedBHProcaLevel::initialData()
+{
+    CH_TIME("BoostedBHProcaLevel::initialData");
+    if (m_verbosity)
+        pout() << "BoostedBHProcaLevel::initialData " << m_level << endl;
+
+    // First set everything to zero, then set the value of the conformal factor
+    // This is just for the diagnostics
+    SetValue set_zero(0.0);
+    BoostedBH boosted_bh(m_p.bg_params, m_dx); // just calculates chi
+    auto compute_pack = make_compute_pack(set_zero, boosted_bh);
+    BoxLoops::loop(compute_pack, m_state_diagnostics, m_state_diagnostics,
+                   SKIP_GHOST_CELLS);
+
+    // Now set the actual evolution variables
+    InitialProcaData initial_pf(m_p.proca_amplitude, m_p.proca_mass, m_p.center,
+                                m_p.bg_params, m_dx);
+    BoxLoops::loop(initial_pf, m_state_new, m_state_new, FILL_GHOST_CELLS);
+
+    // now the gauss constraint
+    fillAllGhosts();
+    ProcaConstraint enforce_constraint(m_p.center, m_p.bg_params,
+                                       m_p.proca_mass, m_dx);
+    BoxLoops::loop(enforce_constraint, m_state_new, m_state_new,
+                   EXCLUDE_GHOST_CELLS);
+
+    // excise evolution vars within horizon, turn off simd vectorisation
+    BoxLoops::loop(
+        ExcisionEvolution<ProcaField, BoostedBH>(m_dx, m_p.center, boosted_bh),
+        m_state_new, m_state_new, SKIP_GHOST_CELLS, disable_simd());
+}
+
+void BoostedBHProcaLevel::specificPostTimeStep()
+{
+    // Check for nans on every level
+    if (m_p.nan_check)
+        BoxLoops::loop(NanCheck(), m_state_new, m_state_new, SKIP_GHOST_CELLS,
+                       disable_simd());
+
+    // At any level, but after the timestep on the minimum extraction level
+    int min_level = m_p.extraction_params.min_extraction_level();
+    bool calculate_diagnostics = at_level_timestep_multiple(min_level);
+    if (calculate_diagnostics)
+    {
+        fillAllGhosts();
+        ProcaField proca_field(m_p.proca_mass, m_p.proca_damping);
+        BoostedBH boosted_bh(m_p.bg_params, m_dx);
+        EnergyConservation<ProcaField, BoostedBH> energies(
+            proca_field, boosted_bh, m_dx, m_p.center);
[…]
