+++ Source/Matter/ProcaField.impl.hpp
@@ -1,0 +1,195 @@
+/* GRChombo
+ * Copyright 2012 The GRChombo collaboration.
+ * Please refer to LICENSE in GRChombo's root directory.
+ */
+
+#if !defined(PROCAFIELD_HPP_)
+#error "This file should only be included through ProcaField.hpp"
+#endif
+
+#ifndef PROCAFIELD_IMPL_HPP_
+#define PROCAFIELD_IMPL_HPP_
+
+// Calculate the stress energy tensor elements
+template <class data_t, template <typename> class vars_t>
+emtensor_t<data_t> ProcaField::compute_emtensor(
+    const vars_t<data_t> &vars, const MetricVars<data_t> &metric_vars,
+    const vars_t<Tensor<1, data_t>> &d1, const Tensor<2, data_t> &gamma_UU,
+    const Tensor<3, data_t> &chris_phys_ULL) const
+{
+    emtensor_t<data_t> out;
+
+    // Some useful quantities
+    const double msquared = pow(m_proca_mass, 2.0);
+
+    // D_i A_j
+    Tensor<2, data_t> DA;
+    FOR2(i, j)
+    {
+        DA[i][j] = d1.Avec[j][i];
+        FOR1(k) { DA[i][j] += -chris_phys_ULL[k][i][j] * vars.Avec[k]; }
+    }
+
+    // D_i A_j - D_j A_i (NB exterior derivative, so christoffel symbols cancel)
+    Tensor<2, data_t> diff_DA;
+    FOR2(i, j) { diff_DA[i][j] = d1.Avec[j][i] - d1.Avec[i][j]; }
+
+    // Calculate components of EM Tensor
+    // S_ij = T_ij
+    FOR2(i, j)
+    {
+        out.Sij[i][j] = msquared * (vars.Avec[i] * vars.Avec[j] +
+                                    0.5 * metric_vars.gamma[i][j] * vars.Avec0 *
+                                        vars.Avec0);
+
+        FOR2(k, l)
+        {
+            out.Sij[i][j] += gamma_UU[k][l] * (diff_DA[i][k] * diff_DA[j][l]) -
+                             metric_vars.gamma[i][k] * metric_vars.gamma[j][l] *
+                                 vars.Evec[k] * vars.Evec[l] +
+                             0.5 * metric_vars.gamma[k][l] *
+                                 metric_vars.gamma[i][j] * vars.Evec[k] *
+                                 vars.Evec[l] -
+                             0.5 * gamma_UU[k][l] * metric_vars.gamma[i][j] *
+                                 msquared * vars.Avec[k] * vars.Avec[l];
+            FOR2(m, n)
+            {
+                out.Sij[i][j] += -0.5 * metric_vars.gamma[i][j] *
+                                 gamma_UU[k][m] * gamma_UU[l][n] * DA[k][l] *
+                                 diff_DA[m][n];
+            }
+        }
+    }
+
+    // S = Tr_S_ij
+    out.S = 0.0;
+    FOR2(i, j) { out.S += out.Sij[i][j] * gamma_UU[i][j]; }
+
+    // S_i (note lower index) = n^a T_a0
+    FOR1(i)
+    {
+        out.Si[i] = msquared * vars.Avec0 * vars.Avec[i];
+
+        FOR1(j) { out.Si[i] += vars.Evec[j] * diff_DA[i][j]; }
+    }
+
+    // rho = n^a n^b T_ab
+    out.rho = 0.5 * msquared * (vars.Avec0 * vars.Avec0);
+    FOR2(i, j)
+    {
+        out.rho +=
[…]
