### 2024-03-07 eca9161
added uniform initial data for Proca field example
