+++ Examples/BoostedBHProca/InitialProcaData.hpp
@@ -48,37 +48,25 @@ class InitialProcaData
         // where am i?
         Coordinates<data_t> coords(current_cell, m_dx, m_center);
 
-        // get the metric vars
-        BoostedBH boosted_bh(m_bg_params, m_dx);
-        MetricVars<data_t> metric_vars;
-        boosted_bh.compute_metric_background(metric_vars, coords);
-        // const auto metric_vars = current_cell.template
-        // load_vars<MetricVars>();
-        const data_t det_gamma =
-            TensorAlgebra::compute_determinant_sym(metric_vars.gamma);
-
-        // useful coordinate quantities
-        const double M = m_bg_params.mass;
-        const double z = coords.z;
-        const data_t r = coords.get_radius();
-        const data_t r2 = r * r;
-
-        // the peak of the bound state for the n=l=0, j=m=1 state
-        // See arXiv : 1704.05081 eqns 9 and 10
-        double alpha = M * m_mu;
-        double r_0 = 1.0 / (m_mu * alpha);
-
-        // set the field variable to approx profile
+        // set the field variables to constant values
         Vars<data_t> vars;
         VarsTools::assign(vars, 0.);
-        data_t Avec = m_amplitude * (exp(-r / r_0)) / det_gamma;
-        double coswt = 1.0;
-        double sinwt = 0.0;
 
         // set the vector values
-        vars.Avec[0] = -Avec * coswt;
-        vars.Avec[1] = -Avec * sinwt;
+        vars.Avec[0] = m_params.proca_amplitude;
+        vars.Avec[1] = 0.0;
         vars.Avec[2] = 0.0;
+        
+        // set auxiliary field values
+        vars.Zvec = 0.0;
+
+        // set electric field values
+        vars.Evec[0] = 0.0;
+        vars.Evec[1] = 0.0;
+        vars.Evec[2] = 0.0;
+
+        // set scalar part of Proca field
+        vars.Avec0 = 0.0;
 
         // Store the initial values of the variables
         current_cell.store_vars(vars);
