### 2024-03-07 b0dc5c0
added uniform initial data for Proca field example
