+++ Examples/BoostedBHProca/InitialProcaData.hpp
@@ -8,6 +8,7 @@
 
 #include "ADMFixedBGVars.hpp"
 #include "BoostedBH.hpp"
+#include "TensorAlgebra.hpp"
 #include "Cell.hpp"
 #include "Coordinates.hpp"
 #include "ProcaField.hpp"
@@ -21,7 +22,6 @@ class InitialProcaData
 {
   protected:
     const double m_amplitude;
-    const double m_mu;
     const double m_dx;
     const std::array<double, CH_SPACEDIM> m_center;
     const BoostedBH::params_t m_bg_params;
@@ -34,10 +34,10 @@ class InitialProcaData
 
   public:
     //! The constructor for the class
-    InitialProcaData(const double a_amplitude, const double a_mu,
+    InitialProcaData(const double a_amplitude, 
                      const std::array<double, CH_SPACEDIM> a_center,
                      const BoostedBH::params_t a_bg_params, const double a_dx)
-        : m_dx(a_dx), m_amplitude(a_amplitude), m_mu(a_mu), m_center(a_center),
+        : m_dx(a_dx), m_amplitude(a_amplitude), m_center(a_center),
           m_bg_params(a_bg_params)
     {
     }
@@ -48,22 +48,35 @@ class InitialProcaData
         // where am i?
         Coordinates<data_t> coords(current_cell, m_dx, m_center);
 
+        // get the metric vars
+        BoostedBH boosted_bh(m_bg_params, m_dx);
+        MetricVars<data_t> metric_vars;
+        boosted_bh.compute_metric_background(metric_vars, coords);
+        
+        //Compute contravariant spatial metric
+        Tensor<2,data_t> gamma_UU = TensorAlgebra::compute_inverse_sym(metric_vars.gamma);
+
         // set the field variables to constant values
         Vars<data_t> vars;
         VarsTools::assign(vars, 0.);
 
         // set the vector values
-        vars.Avec[0] = m_params.proca_amplitude;
+        vars.Avec[0] = m_amplitude;
         vars.Avec[1] = 0.0;
         vars.Avec[2] = 0.0;
         
         // set auxiliary field values
         vars.Zvec = 0.0;
 
-        // set electric field values. Maybe need to set E^i = Avec_j K^{ji} for consistent initial conditions?
-        vars.Evec[0] = 0.0;
-        vars.Evec[1] = 0.0;
-        vars.Evec[2] = 0.0;
+        // set electric field values consistent with Avec and Avec0
+        FOR1(i)
+        {
+          vars.Evec[i] = 0.0;
+          FOR3(j,k,l)
+          {
+            vars.Evec[i] += gamma_UU[i][k] * gamma_UU[j][l] * vars.Avec[j] * metric_vars.K_tensor[k][l];
+          }
+        }
 
         // set scalar part of Proca field
         vars.Avec0 = 0.0;
