+++ Examples/BoostedBHProca/BoostedBHProcaLevel.cpp
@@ -105,7 +105,7 @@ void BoostedBHProcaLevel::specificPostTimeStep()
             double rhoLinMom_sum = amr_reductions.sum(c_rhoLinMom);
             double sourceLinMom_sum = amr_reductions.sum(c_sourceLinMom);
 
-            SmallDataIO integral_file(m_p.data_path + "EnergyIntegrals", m_dt,
+            SmallDataIO integral_file(m_p.data_path + m_p.integrals_filename, m_dt,
                                       m_time, m_restart_time,
                                       SmallDataIO::APPEND, first_step);
             // remove any duplicate data if this is post restart
@@ -117,9 +117,9 @@ void BoostedBHProcaLevel::specificPostTimeStep()
             // write data
             if (first_step)
             {
-                integral_file.write_header_line({"Energy density.",
-                                                 "Lin. Mom. density",
-                                                 "Lin. Mom. source"});
+                integral_file.write_header_line({"Energy",
+                                                 "Linear_Momentum",
+                                                 "Linear_Momentum_Source"});
             }
             integral_file.write_time_data_line(data_for_writing);
 
