+++ Examples/BoostedBHProca/params.txt
@@ -7,7 +7,7 @@
 verbosity = 0
 
 # location / naming of output files
-# output_path = "" # Main path for all files. Must exist!
+output_path = "/pfs/work7/workspace/scratch/hd_pb293-WS_GRChombo/testing/Proca_DynamicalFriction_mu_05_v_03_23217939"
 chk_prefix = BoostedBHProca_
 plot_prefix = BoostedBHProcap_
 # restart_file = BoostedBHProca_000000.3d.hdf5
@@ -32,6 +32,9 @@ print_progress_only_to_rank_0 = 1
 
 # ignore_checkpoint_name_mismatch = 0
 # write_plot_ghosts = 0
+
+#Integrals filename
+integrals_filename = "Integrals"
 
 #################################################
 # Initial Data parameters
