### 2024-03-12 #5956 @warrickball
> I've addressed your comments above in this pull request https://github.com/GRTLCollaboration/GRDzhadzha/pull/21

I've had a look at the paper produced in [the JOSS workflow run](https://github.com/GRTLCollaboration/GRDzhadzha/actions/runs/8141926345?pr=21) and the citations are still a mixed bag of parentheses. e.g.
* *(see e.g. Macedo et al. (2013); ...)* should be *(see e.g. Macedo et al. 2013; ...)*
* *... spacetime metric Clough (2021).* should be *... spacetime metric (Clough 2021).*
* *... NR code GRChombo Clough et al. (2015);Andrade et al. (2021), ...* should be *... NR code GRChombo (Clough et al. 2015; Andrade et al. 2021), ...*.

I know these are tricky in the JOSS source format, so I'll have a go now at resolving these and, presuming I succeed (which is not guaranteed!), I'll open a PR against your branch with the other changes.
