### 2024-03-12 4d7e851
fixes to citation style
