### 2024-03-12 #5956 @warrickball
I've opened https://github.com/GRTLCollaboration/GRDzhadzha/pull/22 to fix *almost* all the citations. The citations in footnote 1 aren't quite right but I'll consult the other JOSS editors about this.

I'll wait until I re-generate the PDF in this issue (after those PRs https://github.com/GRTLCollaboration/GRDzhadzha/pull/21 and https://github.com/GRTLCollaboration/GRDzhadzha/pull/22 are merged) before I tick the other editorial items in my list above but I can see in my local build that you've addressed them. Thanks!
