### 2024-03-12 5cb5aed
fix citations in footnote 1
