### 2024-03-18 3e0a9b2
added more comments to BoostedBHProcaLevel::InitialData
Cleaned up unneccesary debugging code

added initial profile selection flag to parameter file and SimulationParameters.hpp
