### 2024-03-18 #5956 @warrickball
Two catches from the reference check:
* It looks like [Traykova et al. (2023) has been published](https://ui.adsabs.harvard.edu/abs/2023PhRvD.108l1502T/abstract), so you can update the reference, which should include a DOI.
* You should change the *in prep* paper (Wang et al. 2023) to 2024.

I'm not worried about any of the other missing DOIs.

You can then proceed with the items in the [Post-Review Checklist for Editor and Authors](https://github.com/openjournals/joss-reviews/issues/5956#issuecomment-1931651771). In particular, you should create a tagged release and archive and upload it to an archiving service like [Zenodo](https://zenodo.org/) or [Figshare](https://figshare.com/), and post the version number and archive DOI here so I can attach them to our metadata.

You're welcome to do it however you like but when I once published in JOSS, I followed [instructions like these](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content) to publish from GitHub to Zenodo. I've never tried Figshare but [something similar looks possible](https://help.figshare.com/article/how-to-connect-figshare-with-your-github-account).
