### 2024-03-18 84a33fe
Update publication info in some references
