+++ paper/paper.bib
@@ -1145,7 +1145,11 @@ year = {1912},
     eprint = "2305.10492",
     archivePrefix = "arXiv",
     primaryClass = "gr-qc",
-    month = "5",
+    doi = "10.1103/PhysRevD.108.L121502",
+    journal = "Phys. Rev. D",
+    volume = "108",
+    number = "12",
+    pages = "L121502",
     year = "2023"
 }
 
@@ -1177,28 +1181,23 @@ year = {1912},
     year = "2022"
 }
 
-@article{Wang:inprep,
+@article{Wang:2024cej,
     author = "Wang, Zipeng and Helfer, Thomas and Traykova, Dina and Clough, Katy and Berti, Emanuele",
-    title = "{The Magnus Effect on black holes in scalar dark matter environments}",
-    doi = "",
-    journal = "in preparation",
-    volume = "",
-    number = "",
-    pages = "",
-    year = "2023"
-}
-
-@article{Aurrekoetxea:inprep,
-    author = "Aurrekoetxea, Josu C. and Bamber, Jamie and Clough, Katy and Ferreira, Pedro G.",
-    title = "{The effect of wave dark matter on binary black hole mergers}",
-    eprint = "in preparation",
+    title = "{Gravitational Magnus effect from scalar dark matter}",
+    eprint = "2402.07977",
     archivePrefix = "arXiv",
     primaryClass = "gr-qc",
-    doi = "",
-    journal = "",
-    volume = "",
-    number = "",
-    pages = "",
+    month = "2",
+    year = "2024"
+}
+
+@article{Aurrekoetxea:2023jwk,
+    author = "Aurrekoetxea, Josu C. and Clough, Katy and Bamber, Jamie and Ferreira, Pedro G.",
+    title = "{The effect of wave dark matter on equal mass black hole mergers}",
+    eprint = "2311.18156",
+    archivePrefix = "arXiv",
+    primaryClass = "gr-qc",
+    month = "11",
     year = "2023"
 }
 
