### 2024-03-18 e5062b5
Add proca constraint excision within the horizon
