### 2024-03-19 4928f24
added gauss diagnositc tracking
revamped diagnostic excision to automatically account for changes in DiagnosticVariables.hpp

Added black hole horizon calculation to ExcisionProcaEvolution.hpp. Now excising Zvec based of this calculation.
