+++ Examples/BoostedBHProca/DiagnosticVariables.hpp
@@ -15,6 +15,7 @@ enum
     c_rhoLinMom,
     c_fluxLinMom,
     c_sourceLinMom,
+    c_Zvec_out, //Auxiliary Z field. Used for tracking (minus) the Gauss constraint
 
     NUM_DIAGNOSTIC_VARS
 };
@@ -23,7 +24,7 @@ namespace DiagnosticVariables
 {
 static const std::array<std::string, NUM_DIAGNOSTIC_VARS> variable_names = {
     "chi",       "rhoEnergy",  "fluxEnergy",
-    "rhoLinMom", "fluxLinMom", "sourceLinMom"};
+    "rhoLinMom", "fluxLinMom", "sourceLinMom", "Zvec"};
 }
 
 #endif /* DIAGNOSTICVARIABLES_HPP */
