### 2024-03-19 7a3b166
Added further diagnostic classes (see ProcaDiagnostics.hpp) and implemented them in the BoxLoop in specificPostTimestep
added Zvec to diagnostic variables to examine Gauss constraint in post-processing

revamped diagnostic excision to handle changes to DiagnosticVariables.hpp enum (see ExcisionDiagnostics.hpp)
