+++ Examples/BoostedBHProca/BoostedBHProcaLevel.cpp
@@ -29,7 +29,7 @@
 #include "ProcaConstraint.hpp"
 #include "ProcaField.hpp"
 
-//Diagnostics
+// Diagnostics
 #include "ProcaDiagnostic.hpp"
 
 // Initial data for field and metric variables
@@ -59,7 +59,7 @@ void BoostedBHProcaLevel::initialData()
         FILL_GHOST_CELLS); // Loop over box cells, filling the ghost cells
 
     // now the gauss constraint
-    fillAllGhosts(); 
+    fillAllGhosts();
     ProcaConstraint enforce_constraint(
         m_p.center, m_p.bg_params, m_p.proca_mass,
         m_dx); // This sets the scalar part of the Proca field, using the gauss
@@ -98,9 +98,10 @@ void BoostedBHProcaLevel::specificPostTimeStep()
         LinearMomConservation<ProcaField, BoostedBH> linear_momenta(
             proca_field, boosted_bh, direction, m_dx, m_p.center);
 
-        auto compute_pack = make_compute_pack(energies, linear_momenta, gauss_constraint);
-        BoxLoops::loop(compute_pack, m_state_new,
-                       m_state_diagnostics, SKIP_GHOST_CELLS);
+        auto compute_pack =
+            make_compute_pack(energies, linear_momenta, gauss_constraint);
+        BoxLoops::loop(compute_pack, m_state_new, m_state_diagnostics,
+                       SKIP_GHOST_CELLS);
 
         // excise within/outside specified radii, no simd
         BoxLoops::loop(
