### 2024-03-19 e49b150
Fixed initial condition code to satisfy Gauss constraint
