+++ Source/Matter/ComplexProcaField.impl.hpp
@@ -1,0 +1,324 @@
+/* GRChombo
+ * Copyright 2012 The GRChombo collaboration.
+ * Please refer to LICENSE in GRChombo's root directory.
+ */
+
+#if !defined(COMPLEXPROCAFIELD_HPP_)
+#error "This file should only be included through ComplexProcaField.hpp"
+#endif
+
+#ifndef COMPLEXPROCAFIELD_IMPL_HPP_
+#define COMPLEXPROCAFIELD_IMPL_HPP_
+
+// Calculate the stress energy tensor elements
+template <class data_t, template <typename> class vars_t>
+emtensor_t<data_t> ComplexProcaField::compute_emtensor(
+    const vars_t<data_t> &vars, const MetricVars<data_t> &metric_vars,
+    const vars_t<Tensor<1, data_t>> &d1, const Tensor<2, data_t> &gamma_UU,
+    const Tensor<3, data_t> &chris_phys_ULL) const
+{
+    emtensor_t<data_t> out;
+
+    // Some useful quantities
+    const double msquared = pow(m_vector_mass, 2.0);
+
+    // Real parts first then Im
+    // D_i A_j
+    Tensor<2, data_t> DA_Re;
+    Tensor<2, data_t> DA_Im;
+    FOR2(i, j)
+    {
+        DA_Re[i][j] = d1.Avec_Re[j][i];
+        DA_Im[i][j] = d1.Avec_Im[j][i];
+        FOR1(k)
+        {
+            DA_Re[i][j] += -chris_phys_ULL[k][i][j] * vars.Avec_Re[k];
+            DA_Im[i][j] += -chris_phys_ULL[k][i][j] * vars.Avec_Im[k];
+        }
+    }
+
+    // D_i A_j - D_j A_i (NB exterior derivative, so christoffel symbols cancel)
+    Tensor<2, data_t> diff_DA_Re;
+    Tensor<2, data_t> diff_DA_Im;
+    FOR2(i, j)
+    {
+        diff_DA_Re[i][j] = d1.Avec_Re[j][i] - d1.Avec_Re[i][j];
+        diff_DA_Im[i][j] = d1.Avec_Im[j][i] - d1.Avec_Im[i][j];
+    }
+
+    // Calculate components of EM Tensor
+    // S_ij = T_ij
+    FOR2(i, j)
+    {
+        out.Sij[i][j] =
+            // Re part
+            msquared * (vars.Avec_Re[i] * vars.Avec_Re[j] +
+                        0.5 * metric_vars.gamma[i][j] * vars.Avec0_Re *
+                            vars.Avec0_Re) +
+            // Im part
+            msquared *
+                (vars.Avec_Im[i] * vars.Avec_Im[j] +
+                 0.5 * metric_vars.gamma[i][j] * vars.Avec0_Im * vars.Avec0_Im);
+
+        FOR2(k, l)
+        {
+            out.Sij[i][j] +=
+                // Re part
+                gamma_UU[k][l] * (diff_DA_Re[i][k] * diff_DA_Re[j][l]) -
+                metric_vars.gamma[i][k] * metric_vars.gamma[j][l] *
+                    vars.Evec_Re[k] * vars.Evec_Re[l] +
+                0.5 * metric_vars.gamma[k][l] * metric_vars.gamma[i][j] *
+                    vars.Evec_Re[k] * vars.Evec_Re[l] -
+                0.5 * gamma_UU[k][l] * metric_vars.gamma[i][j] * msquared *
+                    vars.Avec_Re[k] * vars.Avec_Re[l] +
+                // Im part
+                gamma_UU[k][l] * (diff_DA_Im[i][k] * diff_DA_Im[j][l]) -
[…]
