+++ Examples/BoostedBHComplexProca/BoostedBHComplexProcaLevel.cpp
@@ -1,0 +1,181 @@
+/* GRChombo
+ * Copyright 2012 The GRChombo collaboration.
+ * Please refer to LICENSE in GRChombo's root directory.
+ */
+
+// General includes common to most GR problems
+#include "BoostedBHComplexProcaLevel.hpp"
+#include "AMRReductions.hpp"
+#include "BoxLoops.hpp"
+#include "ComputePack.hpp"
+#include "NanCheck.hpp"
+#include "SetValue.hpp"
+#include "SmallDataIO.hpp"
+
+// For RHS update
+#include "BoostedBH.hpp"
+#include "MatterEvolution.hpp"
+
+// For tag cells
+#include "FixedGridsTaggingCriterion.hpp"
+
+// Problem specific includes
+#include "EnergyConservation.hpp"
+#include "ExcisionDiagnostics.hpp"
+#include "ExcisionComplexProcaEvolution.hpp"
+#include "FluxExtraction.hpp"
+#include "InitialComplexProcaData.hpp"
+#include "LinearMomConservation.hpp"
+#include "ProcaConstraint.hpp"
+#include "ComplexProcaField.hpp"
+
+// Diagnostics
+#include "ProcaDiagnostic.hpp"
+
+// Initial data for field and metric variables
+void BoostedBHComplexProcaLevel::initialData()
+{
+    CH_TIME("BoostedBHComplexProcaLevel::initialData");
+    if (m_verbosity)
+        pout() << "BoostedBHComplexProcaLevel::initialData " << m_level << endl;
+
+    // First set everything to zero, then set the value of the conformal factor
+    // This is just for the diagnostics
+    SetValue set_zero(0.0);
+    BoostedBH boosted_bh(m_p.bg_params, m_dx); // just calculates chi
+    auto compute_pack = make_compute_pack(
+        set_zero,
+        boosted_bh); // pack up the classes into a single BoxLoops call
+    BoxLoops::loop(
+        compute_pack, m_state_diagnostics, m_state_diagnostics,
+        SKIP_GHOST_CELLS); // Loop over box cells, skipping the ghost cells
+
+    // Now set the actual evolution variables
+    InitialProcaData initial_pf(
+        m_p.proca_amplitude, m_p.proca_mass, m_p.center, m_p.bg_params, m_dx,
+        m_p.proca_initial_data_profile); // Initial proca field data
+    BoxLoops::loop(
+        initial_pf, m_state_new, m_state_new,
+        FILL_GHOST_CELLS); // Loop over box cells, filling the ghost cells
+
+    // now the gauss constraint
+    fillAllGhosts();
+    ProcaConstraint enforce_constraint(
+        m_p.center, m_p.bg_params, m_p.proca_mass,
+        m_dx); // This sets the scalar part of the Proca field, using the gauss
+               // constraint
+    BoxLoops::loop(
+        enforce_constraint, m_state_new, m_state_new,
+        EXCLUDE_GHOST_CELLS); // Loop over box cells, excluding the ghost cells
+
+    // excise evolution vars within horizon, turn off simd vectorisation
+    BoxLoops::loop(ExcisionProcaEvolution<ComplexProcaField, BoostedBH>(
+                       m_dx, m_p.center, boosted_bh),
+                   m_state_new, m_state_new, SKIP_GHOST_CELLS,
+                   disable_simd()); // Now excise the evolution variables inside
+                                    // the horizon
+}
+
+void BoostedBHComplexProcaLevel::specificPostTimeStep()
+{
+    // Check for nans on every level
+    if (m_p.nan_check)
+        BoxLoops::loop(NanCheck(), m_state_new, m_state_new, SKIP_GHOST_CELLS,
+                       disable_simd());
+
[…]
