+++ Examples/BoostedBHProca/ExcisionDiagnostics.hpp
@@ -40,18 +40,11 @@ template <class matter_t, class background_t> class ExcisionDiagnostics
         const Coordinates<double> coords(current_cell, m_dx, m_center);
         if (coords.get_radius() < m_inner_r || coords.get_radius() > m_outer_r)
         {
-            /* current_cell.store_vars(0.0, c_rhoLinMom);
+            // We only want the densities and sources ecised for volume integral
+            // calculateions
+            current_cell.store_vars(0.0, c_rhoLinMom);
             current_cell.store_vars(0.0, c_rhoEnergy);
-            current_cell.store_vars(0.0, c_sourceLinMom); */
-
-            // We do a little bit fancier diagnostic excision here, using the
-            // fact that enum's are just integers This approach automatically
-            // takes into account any changes to the number of diagnostic
-            // variables in DiagnosticVariables.hpp
-            for (int i{0}; i < NUM_DIAGNOSTIC_VARS; i++)
-            {
-                current_cell.store_vars(0.0, i);
-            };
+            current_cell.store_vars(0.0, c_sourceLinMom);
         }
     }
 };
