### 2024-04-08 807dac7
Fix syntax error in pair of references in paper
