### 2024-04-08 #5956 @warrickball
Hi again @dinatraykova, I have to sheepishly admit I introduced a syntax error in my fix (citations should be separated by `;`, not `,`) and opened [a new PR](https://github.com/GRTLCollaboration/GRDzhadzha/pull/25) to fix that. Apologies for slowing things down with this slip.
